More tests

+++ EnvueStreamSelection/EnvueStreamSelection/Broadcast/BroadcastCollectionExtensions.cs
@@ -5,13 +5,13 @@ namespace EnvueStreamSelection
 {
     public static class BroadcastCollectionExtensions
     {
-        public static Dictionary<IBroadcast, int> GetNormalisedBitrates(this ICollection<IBroadcast> broadcasts)
+        public static Dictionary<IBroadcast, double> GetNormalisedBitrates(this ICollection<IBroadcast> broadcasts)
         {
             if (broadcasts.Count == 0)
-                return new Dictionary<IBroadcast, int>();
+                return new Dictionary<IBroadcast, double>();
             
             var maxBitrate = broadcasts.Max(e => e.Bitrate);
-            return broadcasts.ToDictionary(b => b, b => maxBitrate > 0 ? b.Bitrate / maxBitrate : 0);
+            return broadcasts.ToDictionary(b => b, b => maxBitrate > 0 ? b.Bitrate / (double) maxBitrate : 0D);
         }
     }
 }
